+++ lib/eve_online/esi/models/constellation.rb
@@ -8,9 +8,6 @@ module EveOnline
           {
             constellation_id: constellation_id,
             name: name,
-            position_x: position_x,
-            position_y: position_y,
-            position_z: position_z,
             region_id: region_id,
             systems: systems
           }
@@ -24,24 +21,16 @@ module EveOnline
           options['name']
         end
 
-        def position_x
-          options['position']['x'] if options['position']
-        end
-
-        def position_y
-          options['position']['y'] if options['position']
-        end
-
-        def position_z
-          options['position']['z'] if options['position']
-        end
-
         def region_id
           options['region_id']
         end
 
         def systems
           options['systems']
+        end
+
+        def position
+          @position ||= Position.new(options['position'])
         end
       end
     end